Refactor item logic into item partial
